worker/requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
google-generativeai==0.4.0
google-api-python-client==2.116.0
google-auth==2.27.0
PyMuPDF==1.23.8
pgvector==0.2.5
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0
